backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
PyMuPDF==1.23.26
python-docx==1.1.0
tantivy==0.21.0
numpy==1.26.4
openai==1.14.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
